requirements.txt
mysql-connector
networkx
numpy
matplotlib
phpserialize